@@ -2,7 +2,7 @@ package com.example.jaerencoathup.exampleapp.activities;
 
 import android.app.AlertDialog;
 import android.app.ProgressDialog;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 
 import javax.inject.Inject;
 
