@@ -1,9 +1,9 @@
 package com.example.jaerencoathup.exampleapp.persistence.dao;
 
-import android.arch.persistence.room.Dao;
-import android.arch.persistence.room.Insert;
-import android.arch.persistence.room.OnConflictStrategy;
-import android.arch.persistence.room.Query;
+import androidx.room.Dao;
+import androidx.room.Insert;
+import androidx.room.OnConflictStrategy;
+import androidx.room.Query;
 import com.example.jaerencoathup.exampleapp.persistence.WeatherData;
 
 import io.reactivex.Maybe;
