@@ -1,6 +1,6 @@
 package com.example.jaerencoathup.exampleapp.persistence.typeconverters;
 
-import android.arch.persistence.room.TypeConverter;
+import androidx.room.TypeConverter;
 
 import org.joda.time.DateTime;
 
